pyproject.toml
[project]
name = "CardioLogAI"
version = "0.1.0"
description = "A information system based on machine learning to help in the early diagnosis of cardivascular disease"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "sqlmodel>=0.0.24",
]

[tool.poetry.dependencies]
fastapi = "^0.116.1"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.5.0"
pydantic = "^2.0.0"
sqlmodel = "^0.0.24"
psycopg2-binary = "^2.9.9"

[tool.poetry.dev-dependencies]
pytest = "^8.2.0"
httpx = "^0.27.0"
